# albums-related/album.controller.ts
import {
    Body,
    Controller,
    Delete,
    Get,
    HttpCode,
    Param,
    Post,
    Put,
} from '@nestjs/common';
import { AlbumService } from './album.service';
import {CreateAlbumDto} from "./dto/create-album.dto";

@Controller('album')
export class AlbumController {
    constructor(private readonly albumService: AlbumService) {}

    @Get()
    getAlbums() {
        return this.albumService.getAlbums();
    }

    @Get(':id')
    getAlbum(@Param('id') id: string) {
        return this.albumService.getAlbumById(id);
    }

    @Post()
    @HttpCode(201)
    create(@Body() albumDto: CreateAlbumDto) {
        return this.albumService.createAlbum(albumDto);
    }

    @Put(':id')
    @HttpCode(200)
    updateAlbum(@Param('id') id: string, @Body() updateAlbumDto: CreateAlbumDto) {
        return this.albumService.updateAlbum(id, updateAlbumDto);
    }

    @Delete(':id')
    @HttpCode(204)
    deleteAlbum(@Param('id') id: string) {
        return this.albumService.deleteAlbum(id);
    }
}